# Add implementation of negativePatchPoint.