--- a/src/com/amazon/ion/impl/bin/IonRawBinaryWriter.java
+++ b/src/com/amazon/ion/impl/bin/IonRawBinaryWriter.java
@@ -771,7 +771,7 @@ import java.util.NoSuchElementException;
                 // We've reclaimed some number of bytes; adjust the container length as appropriate.
                 length -= numberOfBytesToShiftBy;
             }
-            else if (currentContainer.length <= preallocationMode.contentMaxLength)
+            else if (length <= preallocationMode.contentMaxLength)
             {
                 // The container's encoded body is too long to fit the length in the type descriptor byte, but it will
                 // fit in the preallocated length bytes that were added to the buffer when the container was started.