@linlin-s 2023-05-23
This is the benchmark result of the new commit:
```
Benchmark                                                                   (input)                                       (options)  Mode  Cnt          Score           Error   Units
Bench.run                                   /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10       3974.017 ±        65.594   ms/op
Bench.run:Heap usage                        /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10       3416.969 ±       305.973      MB
Bench.run:Serialized size                   /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10        201.663                      MB
Bench.run:·gc.alloc.rate                    /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10        215.062 ±         5.421  MB/sec
Bench.run:·gc.alloc.rate.norm               /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10  937919562.400 ±     11171.238    B/op
Bench.run:·gc.churn.G1_Eden_Space           /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10         89.568 ±         8.110  MB/sec
Bench.run:·gc.churn.G1_Eden_Space.norm      /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10  390699417.600 ±  36179338.066    B/op
Bench.run:·gc.churn.G1_Old_Gen              /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10         47.603 ±        26.417  MB/sec
Bench.run:·gc.churn.G1_Old_Gen.norm         /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10  206884044.800 ± 112060198.739    B/op
Bench.run:·gc.churn.G1_Survivor_Space       /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10          0.836 ±         1.317  MB/sec
Bench.run:·gc.churn.G1_Survivor_Space.norm  /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10    3635063.467 ±   5681570.222    B/op
Bench.run:·gc.count                         /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10         53.000                  counts
Bench.run:·gc.time                          /home/linls/testData/sample_log_dub.ion  write::{f:ION_BINARY,t:BUFFER,L:2,a:STREAMING}  avgt   10      13971.000                      ms

```