@tgregg 2023-05-23
We can see that the heap usage and GC churn increased significantly with the change, due to allocating significantly more PatchPoint instances. This likely negatively affected the execution time. I am curious what this would look like if we pooled PatchPoint instances, similar to what we do with ContainerInfo using _Private_RecyclingStack. That way PatchPoint instances will only be held until the data is flushed, and will be reused. Is this something you'd be interested in experimenting with?